[MOVE/#305] 컴포넌트 위치 수정
