+++ feature/calendar/src/main/java/com/terning/feature/calendar/calendar/CalendarViewModel.kt
@@ -2,7 +2,7 @@ package com.terning.feature.calendar.calendar
 
 import androidx.lifecycle.ViewModel
 import com.terning.feature.calendar.calendar.model.DayModel
-import com.terning.feature.calendar.calendar.state.CalendarUiState
+import com.terning.feature.calendar.calendar.model.CalendarUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
