+++ feature/calendar/src/main/java/com/terning/feature/calendar/calendar/model/CalendarUiState.kt
@@ -1,9 +1,7 @@
-package com.terning.feature.calendar.calendar.state
+package com.terning.feature.calendar.calendar.model
 
 import androidx.compose.foundation.pager.PagerState
 import androidx.compose.runtime.compositionLocalOf
-import com.terning.feature.calendar.calendar.model.DayModel
-import com.terning.feature.calendar.calendar.model.TerningCalendarModel
 
 data class CalendarUiState(
     val selectedDate: DayModel = DayModel(),
